extracted meta namespace to its own header
