+++ BankOCR/BankOCR.vcxproj.filters
@@ -2,6 +2,7 @@
 <Project ToolsVersion="4.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
   <ItemGroup>
     <ClInclude Include="OCR.h" />
+    <ClInclude Include="meta.h" />
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="OCR.cpp" />
